requirements.txt
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
PyYAML==5.4.1
